Lift startViewTransition implementation to react-router-dom
