+++ .changeset/lift-start-view-transition.md
@@ -1,0 +1,6 @@
+---
+"react-router-dom": patch
+"react-router": patch
+---
+
+Export a separate `RouterProvider` from `react-router-dom` with `startViewTransition` support
