+++ .changeset/start-view-transition.md
@@ -1,6 +1,5 @@
 ---
 "react-router-dom": minor
-"react-router": minor
 "@remix-run/router": minor
 ---
